Fixed mobile search focus - keyboard collapse/focus
