+++ webapp/components/search_bar.jsx
@@ -1,7 +1,6 @@
 // Copyright (c) 2015 Mattermost, Inc. All Rights Reserved.
 // See License.txt for license information.
 
-import $ from 'jquery';
 import * as GlobalActions from 'actions/global_actions.jsx';
 import SearchStore from 'stores/search_store.jsx';
 import UserStore from 'stores/user_store.jsx';
@@ -31,7 +30,9 @@ export default class SearchBar extends React.Component {
         this.handleUserFocus = this.handleUserFocus.bind(this);
         this.handleClear = this.handleClear.bind(this);
         this.handleUserBlur = this.handleUserBlur.bind(this);
-        this.performSearch = this.performSearch.bind(this);
+        this.handleSearch = this.handleSearch.bind(this);
+        this.handleSearchOnSuccess = this.handleSearchOnSuccess.bind(this);
+        this.handleSearchOnError = this.handleSearchOnError.bind(this);
         this.handleSubmit = this.handleSubmit.bind(this);
         this.searchMentions = this.searchMentions.bind(this);
         this.getFlagged = this.getFlagged.bind(this);
@@ -125,7 +126,7 @@ export default class SearchBar extends React.Component {
         this.setState({focused: true});
     }
 
-    performSearch(terms, isMentionSearch) {
+    handleSearch(terms, isMentionSearch) {
         if (terms.length) {
             this.setState({
                 isSearching: true,
@@ -136,23 +137,35 @@ export default class SearchBar extends React.Component {
                 terms,
                 isMentionSearch,
                 () => {
-                    this.setState({isSearching: false});
-
-                    if (Utils.isMobile() && this.search) {
-                        this.search.value = '';
-                    }
+                    this.handleSearchOnSuccess();
                 },
                 () => {
-                    this.setState({isSearching: false});
+                    this.handleSearchOnError();
                 }
             );
         }
     }
 
+    handleSearchOnSuccess() {
+        if (this.mounted) {
+            this.setState({isSearching: false});
+
+            if (Utils.isMobile() && this.search) {
+                this.search.value = '';
+            }
+        }
+    }
+
+    handleSearchOnError() {
+        if (this.mounted) {
+            this.setState({isSearching: false});
+        }
+    }
+
     handleSubmit(e) {
         e.preventDefault();
-        this.performSearch(this.state.searchTerm.trim());
-        $(this.search).find('input').blur();
+        this.handleSearch(this.state.searchTerm.trim());
+        this.search.blur();
     }
 
     searchMentions(e) {
@@ -307,7 +320,7 @@ export default class SearchBar extends React.Component {
                         listComponent={SearchSuggestionList}
                         providers={this.suggestionProviders}
                         type='search'
-                        autoFocus={this.props.isFocus}
+                        autoFocus={this.props.isFocus && this.state.searchTerm === ''}
                     />
                     <div
                         className={clearClass}
